Renamed top-level module to 'golemgpt'
