+++ golemgpt/golems/director.py
