+++ golemgpt/utils/console.py
