+++ golemgpt/utils/images/__main__.py
