+++ golemgpt/utils/images/prompt.py
