+++ golemgpt/utils/images/settings.py
