+++ golemgpt/utils/memory/base.py
